Move visit methods into a Ripper Compiler
